+++ packages/terminal/src/browser/terminal-frontend-contribution.ts
@@ -93,6 +93,33 @@ export namespace TerminalCommands {
         category: TERMINAL_CATEGORY,
         label: 'Hide find widget'
     };
+
+    export const SCROLL_LINE_UP: Command = {
+        id: 'terminal:scroll:line:up',
+        category: TERMINAL_CATEGORY,
+        label: 'Scroll line up'
+    };
+    export const SCROLL_LINE_DOWN: Command = {
+        id: 'terminal:scroll:line:down',
+        category: TERMINAL_CATEGORY,
+        label: 'Scroll line down'
+    };
+    export const SCROLL_TO_TOP: Command = {
+        id: 'terminal:scroll:top',
+        category: TERMINAL_CATEGORY,
+        label: 'Scroll to top'
+    };
+    export const SCROLL_PAGE_UP: Command = {
+        id: 'terminal:scroll:page:up',
+        category: TERMINAL_CATEGORY,
+        label: 'Scroll page up'
+    };
+    export const SCROLL_PAGE_DOWN: Command = {
+        id: 'terminal:scroll:page:down',
+        category: TERMINAL_CATEGORY,
+        label: 'Scroll page down'
+    };
+
     /**
      * Command that displays all terminals that are currently opened
      */
@@ -244,6 +271,42 @@ export class TerminalFrontendContribution implements TerminalService, CommandCon
                 terminalSearchBox.hide();
             }
         });
+
+        commands.registerCommand(TerminalCommands.SCROLL_LINE_UP, {
+                isEnabled: () => this.shell.activeWidget instanceof TerminalWidget,
+                isVisible: () => false,
+                execute: () => {
+                    (this.shell.activeWidget as TerminalWidget).scrollLineUp();
+                }
+        });
+        commands.registerCommand(TerminalCommands.SCROLL_LINE_DOWN, {
+            isEnabled: () => this.shell.activeWidget instanceof TerminalWidget,
+            isVisible: () => false,
+            execute: () => {
+                (this.shell.activeWidget as TerminalWidget).scrollLineDown();
+            }
+        });
+        commands.registerCommand(TerminalCommands.SCROLL_TO_TOP, {
+            isEnabled: () => this.shell.activeWidget instanceof TerminalWidget,
+            isVisible: () => false,
+            execute: () => {
+                (this.shell.activeWidget as TerminalWidget).scrollToTop();
+            }
+        });
+        commands.registerCommand(TerminalCommands.SCROLL_PAGE_UP, {
+            isEnabled: () => this.shell.activeWidget instanceof TerminalWidget,
+            isVisible: () => false,
+            execute: () => {
+                (this.shell.activeWidget as TerminalWidget).scrollPageUp();
+            }
+        });
+        commands.registerCommand(TerminalCommands.SCROLL_PAGE_DOWN, {
+            isEnabled: () => this.shell.activeWidget instanceof TerminalWidget,
+            isVisible: () => false,
+            execute: () => {
+                (this.shell.activeWidget as TerminalWidget).scrollPageDown();
+            }
+        });
     }
 
     registerMenus(menus: MenuModelRegistry): void {
@@ -291,11 +354,35 @@ export class TerminalFrontendContribution implements TerminalService, CommandCon
             keybinding: 'ctrlcmd+f',
             context: TerminalKeybindingContexts.terminalActive
         });
-
         keybindings.registerKeybinding({
             command: TerminalCommands.TERMINAL_FIND_TEXT_CANCEL.id,
             keybinding: 'esc',
             context: TerminalKeybindingContexts.terminalHideSearch
+        });
+        keybindings.registerKeybinding({
+            command: TerminalCommands.SCROLL_LINE_UP.id,
+            keybinding: 'ctrl+shift+up',
+            context: TerminalKeybindingContexts.terminalActive
+        });
+        keybindings.registerKeybinding({
+            command: TerminalCommands.SCROLL_LINE_DOWN.id,
+            keybinding: 'ctrl+shift+down',
+            context: TerminalKeybindingContexts.terminalActive
+        });
+        keybindings.registerKeybinding({
+            command: TerminalCommands.SCROLL_TO_TOP.id,
+            keybinding: 'home',
+            context: TerminalKeybindingContexts.terminalActive
+        });
+        keybindings.registerKeybinding({
+            command: TerminalCommands.SCROLL_PAGE_UP.id,
+            keybinding: 'pageUp',
+            context: TerminalKeybindingContexts.terminalActive
+        });
+        keybindings.registerKeybinding({
+            command: TerminalCommands.SCROLL_PAGE_DOWN.id,
+            keybinding: 'pageDown',
+            context: TerminalKeybindingContexts.terminalActive
         });
 
         /* Register passthrough keybindings for combinations recognized by